refactor: 모달 구조 리팩터링
footer 부분을 body 에서 사용할 수 있도록 구성
