+++ src/components/modals/index.js
@@ -1,0 +1,2 @@
+export { default as ModalFooter } from './ModalFooter';
+export { default as Modals } from './Modals';
